fix(InputView): add a method to validate the given user input is null or empty
Add a new method 'validateCareNames()' to validate the 'List<String>' from 'parseCarNames()', because the program was unable to detect the empty list from user's empty input.

Problem: Program is unable to detect empty list from user's empty input.
Solution: Add new method to detect and validate the list is null or empty, and throw exception if it is.